Move graphics processing to backend
Create a virtualenv and `pip install python_requirements.txt`, then
`python async_tasks/gen_chunk.py 1600 16 4000` (make sure to do it from
project root dir). This will create a folder `public/img/chunks/` and
generate 1600 1600x1600 jpg images, one for each world chunk. Now, when
the client requests a chunk, it does so from the server and the server
sends the image insantaneously over (as opposed to generating in
real-time on client).

Graphics files have been split into `managers.js` and `objects.js`
instead of `base.js`. This is the start of cleanup/modularization.

`http://localhost:3000` is no longer hardcoded and the socket connection
will be made regardless of hostname.

Drawing has been optimized to redraw only moving parts and now uses
`requestAnimationFrame` to request a draw from the CPU.
