+++ .gitignore
@@ -2,3 +2,4 @@
 .DS_Store
 node_modules
 Cappies\ Project\ Proposal\ Report.pdf
+public/img/chunks
