+++ public/js/client.js
@@ -1,72 +1,51 @@
+import {GameManager, AssetManager} from './graphics/managers.js';
+import {Player} from './graphics/objects.js';
+
 document.addEventListener("DOMContentLoaded", function(event) {
-		canvas = document.getElementById('game');
-		canvas.width = window.innerWidth;
-		canvas.height = window.innerHeight;
-		ctx = canvas.getContext('2d');
+		let player = new Player('player');
+		let manager = null;
 
-		player = new Player('player');
-		manager = null;
+		let canvas = document.getElementById('game');
+		canvas.width = window.innerWidth, canvas.height = window.innerHeight;
+		let ctx = canvas.getContext('2d');
 
-		asset_manager = new AssetManager();
-		img_list = ['img/Grass.png', 'img/Sand.png', 'img/Snow.png',
-								'img/Swamp.png', 'img/Jungle.png', 'img/Water.png'];
-		player_stand_n = ['img/Player/standN.png'];
-		player_stand_s = ['img/Player/standS.png'];
-		player_stand_w = ['img/Player/standW.png'];
-		player_stand_e = ['img/Player/standE.png'];
-		player_walk_n = ['img/Player/walk1N.png','img/Player/walk2N.png'];
-		player_walk_s = ['img/Player/walk1S.png','img/Player/walk2S.png'];
-		player_walk_w = ['img/Player/walk1W.png','img/Player/walk2W.png'];
-		player_walk_e = ['img/Player/walk1E.png','img/Player/walk2E.png'];
+	  // Make WebSocket connection
+    let ws = new WebSocket('ws://' + window.location.hostname + ':' + window.location.port);
+    ws.onopen = () => { console.log("Connected to server!"); }
+
+		let asset_manager = new AssetManager();
+		let img_list = [];
+		let player_stand_n = ['img/Player/standN.png'];
+		let player_stand_s = ['img/Player/standS.png'];
+		let player_stand_w = ['img/Player/standW.png'];
+		let player_stand_e = ['img/Player/standE.png'];
+		let player_walk_n = ['img/Player/walk1N.png','img/Player/walk2N.png'];
+		let player_walk_s = ['img/Player/walk1S.png','img/Player/walk2S.png'];
+		let player_walk_w = ['img/Player/walk1W.png','img/Player/walk2W.png'];
+		let player_walk_e = ['img/Player/walk1E.png','img/Player/walk2E.png'];
 		img_list = img_list.concat(player_stand_n, player_stand_s, player_stand_w,
 															 player_stand_e, player_walk_n, player_walk_s,
 															 player_walk_w, player_walk_e);
 
 		asset_manager.load_images(img_list, () => {
-			let biomes = {};
-			biomes['grass'] = new StaticObject(0);
-			biomes['grass'].load_sprite('img/Grass.png');
-			biomes['sand'] = new StaticObject(0);
-			biomes['sand'].load_sprite('img/Sand.png');
-			biomes['snow'] = new StaticObject(0);
-			biomes['snow'].load_sprite('img/Snow.png');
-			biomes['swamp'] = new StaticObject(0);
-			biomes['swamp'].load_sprite('img/Swamp.png');
-			biomes['jungle'] = new StaticObject(0);
-			biomes['jungle'].load_sprite('img/Jungle.png');
-			biomes['water'] = new StaticObject(0);
-			biomes['water'].load_sprite('img/Water.png');
-
 			player.load_animation('idle_n', player_stand_n, 0);
 			player.load_animation('idle_s', player_stand_s, 0);
 			player.load_animation('idle_w', player_stand_w, 0);
 			player.load_animation('idle_e', player_stand_e, 0);
-			player.load_animation('walk_n', player_walk_n, 5);
-			player.load_animation('walk_s', player_walk_s, 5);
-			player.load_animation('walk_w', player_walk_w, 5);
-			player.load_animation('walk_e', player_walk_e, 5);
+			player.load_animation('walk_n', player_walk_n, 1);
+			player.load_animation('walk_s', player_walk_s, 1);
+			player.load_animation('walk_w', player_walk_w, 1);
+			player.load_animation('walk_e', player_walk_e, 1);
 			player.x = 32000;
 			player.y = 32000;
 
-			$.getJSON("js/graphics/gen2k.json", function(gen_json) {
-				console.log('hello')
-				manager = new GameManager(ctx, canvas.width, canvas.height, gen_json, biomes, player);
-				manager.gen_around_player();
-				manager.start();
-			});
+			manager = new GameManager(ctx, canvas.width, canvas.height, player);
+			manager.gen_around_player();
+			manager.start();
 		});
 
-    // Make client id
-    var client_id = -1;
 
-    // Make WebSocket connection
-    var ws = new WebSocket('ws://localhost:3000');
-
-    // When connection is made
-    ws.onopen = () => {
-        console.log("Connected to server!");
-    }
-
+		// click on chat history to hide/unhide
     let chat_hidden = true;
     $('#chat-history').click(function() {
         if (!chat_hidden) {
@@ -77,11 +56,12 @@ document.addEventListener("DOMContentLoaded", function(event) {
         chat_hidden = !chat_hidden;
     });
 
+		// chat input send
     $('form').submit(() => {
         msg = $('#input_message').val();
         var data = {
             type: "chat",
-            data: {player: client_id,
+            data: {player: player.id,
                    text: msg}
         };
         ws.send(JSON.stringify(data));
@@ -89,7 +69,8 @@ document.addEventListener("DOMContentLoaded", function(event) {
         return false;
     });
 
-    var key_map = {
+		// player movement send
+    let key_map = {
         37: false,
         38: false,
         39: false,
@@ -114,10 +95,9 @@ document.addEventListener("DOMContentLoaded", function(event) {
     });
 
 
-
     // Handle message passed from server
     ws.onmessage = (msg) => {
-        var message = {
+        let message = {
             type: "none"
         };
         try {
@@ -128,22 +108,24 @@ document.addEventListener("DOMContentLoaded", function(event) {
         }
 
         if (message.type == "chat") {
-            var chatmsg = message.data.text;
+            let chatmsg = message.data.text;
             message.data.player++;
             $('#chat-history').append($('<li>').text("Player_" + message.data.player + ": " + chatmsg));
         } else if (message.type == "disconnect") {
-            var player_to_drop = message.data.player;
+            let player_to_drop = message.data.player;
 						console.log(player_to_drop)
             manager.drop_object('player_'+player_to_drop);
-        } else if (message.type == "world_data") {
+        }
+			  else if (message.type == "world_data") {
             if (manager != null) {
-                player_keys = Object.keys(message.data);
-                num_of_players = player_keys.length;
+                let player_keys = Object.keys(message.data);
+                let num_of_players = player_keys.length;
                 for (let i = 0; i < num_of_players; i++) {
                     let id = 'player_' + player_keys[i]
-                    if (player_keys[i] == client_id) {
-                        player.x = message["data"][client_id]["x_position"];
-                        player.y = message["data"][client_id]["y_position"];
+										let int_id = parseInt(player_keys[i])
+                    if (id == player.id) {
+                        player.x = message.data[int_id].x_position;
+                        player.y = message.data[int_id].y_position;
                     } else {
                         if (manager._objects[id] == null) {
                             let new_player = new Player(id);
@@ -151,23 +133,22 @@ document.addEventListener("DOMContentLoaded", function(event) {
                             new_player.load_animation('idle_s', player_stand_s, 0);
                             new_player.load_animation('idle_w', player_stand_w, 0);
                             new_player.load_animation('idle_e', player_stand_e, 0);
-                            new_player.load_animation('walk_n', player_walk_n, 5);
-                            new_player.load_animation('walk_s', player_walk_s, 5);
-                            new_player.load_animation('walk_w', player_walk_w, 5);
-                            new_player.load_animation('walk_e', player_walk_e, 5);
-                            new_player.x = message["data"][player_keys[i]]["x_position"];
-                            new_player.y = message["data"][player_keys[i]]["y_position"];
+                            new_player.load_animation('walk_n', player_walk_n, 1);
+                            new_player.load_animation('walk_s', player_walk_s, 1);
+                            new_player.load_animation('walk_w', player_walk_w, 1);
+                            new_player.load_animation('walk_e', player_walk_e, 1);
+                            new_player.x = message.data[int_id].x_position;
+                            new_player.y = message.data[int_id].y_position;
                             manager.add_object(new_player);
                         } else {
-                            manager._objects[id].x = message["data"][player_keys[i]]["x_position"];
-                            manager._objects[id].y = message["data"][player_keys[i]]["y_position"];
+                            manager._objects[id].x = message.data[int_id].x_position;
+                            manager._objects[id].y = message.data[int_id].y_position;
                         }
                     }
                 }
             }
         } else if (message.type == "id") {
-            client_id = message.data.player_id;
-            player.id = 'player_' + client_id;
+            player.id = 'player_' + message.data.player_id;
         }
     };
