+++ public/js/graphics/constants.js
@@ -1,0 +1,7 @@
+export const WORLD_UNIT = 16;  // in px
+export const WORLD_SIZE = 4000;  // in world units (so actual world size in px is 4000*16)
+export const CHUNK_SIZE = 1600;  // in px
+export const FPS = 1000/24;  // in millis
+
+// globally cached sprites
+export var cached_assets = {};
