+++ python_requirements.txt
@@ -1,0 +1,1 @@
+Pillow==5.1.0
