+++ server.js
@@ -4,6 +4,7 @@ const app = express();
 const http = require('http');
 const url = require('url');
 const WebSocket = require('ws');
+const { spawn } = require('child_process');
 
 var port = process.env.PORT || 3000;
 const server = http.createServer(app);
