+++ views/index.ejs
@@ -4,10 +4,9 @@
 	<meta http-equiv="content-type" content="text/html; charset=utf-8" />
 	<title>Cappie's MMO</title>
 	<script src="js/jquery.js"></script>
-	<script src="js/client.js"></script>
+	<script type="module" src="js/client.js"></script>
 	<link rel="stylesheet" href="css/style.css">
 	<script type="text/javascript" src="js/graphics/gen2k.json"></script>
-	<script type="text/javascript" src="js/graphics/base.js"></script>
 </head>
 <body>
 	<canvas id="game"></canvas>
